Fix hit-class Int format with Int64

+++ Analyzer/EuroOpt/Engine/WeightSweepEngine.swift
@@ -411,12 +411,12 @@ final class WeightSweepEngine {
         let expectedZeroShare = totalTickets > 0 ? expected.expectedCount(mainHits: 0, euroHits: 0) / Double(totalTickets) * 100.0 : 0
 
         print("")
-        print(String(format: "Gesamttipps         : %d", Int32(totalTickets)))
-        print(String(format: "Mind. 1 Treffer     : %d (%.2f %%) | Zufall %.2f %% | Δ %+0.2f %-Pkt.",
-                     Int32(winningTickets), winningShare, expectedWinningShare, winningShare - expectedWinningShare))
-        print(String(format: "0 + 0               : %d (%.2f %%) | Zufall %.2f %% | Δ %+0.2f %-Pkt.",
-                     Int32(zeroTickets), zeroShare, expectedZeroShare, zeroShare - expectedZeroShare))
-        print(String(format: "Klassen-Summe       : %d | Zufall %.2f", Int32(aggregate.total), expected.total))
+        print(String(format: "Gesamttipps         : %lld", Int64(totalTickets)))
+        print(String(format: "Mind. 1 Treffer     : %lld (%.2f %%) | Zufall %.2f %% | Δ %+0.2f %-Pkt.",
+                     Int64(winningTickets), winningShare, expectedWinningShare, winningShare - expectedWinningShare))
+        print(String(format: "0 + 0               : %lld (%.2f %%) | Zufall %.2f %% | Δ %+0.2f %-Pkt.",
+                     Int64(zeroTickets), zeroShare, expectedZeroShare, zeroShare - expectedZeroShare))
+        print(String(format: "Klassen-Summe       : %lld | Zufall %.2f", Int64(aggregate.total), expected.total))
         print("")
     }
